refactor(edgeclaw): split installer into orchestrator + per-backend modules
install/install.ts becomes a thin orchestrator that runs EdgeClaw-wide
setup (workspace copy, telegram streaming-off, gateway restart) and
calls each backend installer:

- install_index.ts — Index Network MCP config + cron jobs (extracted)
- install_edgeos.ts — placeholder (calendar + directory)
- install_geo.ts — placeholder (knowledge graph)

Each placeholder header points future contributors at two contribution
surfaces: per-backend OpenClaw config in the install_<backend>.ts file,
and runtime guidance in workspace/ markdown (TOOLS.md, HEARTBEAT.md,
AGENTS.md, etc.).

User-facing command (`bun install/install.ts <API_KEY>`) is unchanged.
package.json files field broadened to include all install/*.ts.
